split list diff - Percent diff field (#2838)
